race_01/pyproject.toml
[project]
name = "race_01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flytekit>=1.16.5",
    "flytekitplugins-deck-standard>=1.16.5",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "xgboost>=3.0.5",
]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "ruff>=0.14.1",
]
